+++ datanode/gateway/graphql/update_referral_program_resolver.go
@@ -28,8 +28,9 @@ func (u updateReferralProgramResolver) BenefitTiers(_ context.Context, obj *vega
 	return obj.Changes.BenefitTiers, nil
 }
 
-func (u updateReferralProgramResolver) EndOfProgramTimestamp(_ context.Context, obj *vega.UpdateReferralProgram) (string, error) {
-	return time.Unix(obj.Changes.EndOfProgramTimestamp, 0).String(), nil
+func (u updateReferralProgramResolver) EndOfProgramTimestamp(_ context.Context, obj *vega.UpdateReferralProgram) (int64, error) {
+	endTime := time.Unix(obj.Changes.EndOfProgramTimestamp, 0)
+	return endTime.UnixNano(), nil
 }
 
 func (u updateReferralProgramResolver) WindowLength(_ context.Context, obj *vega.UpdateReferralProgram) (int, error) {